allow user to save preset after installing
